+++ internal/libtorch/libtorch_cuda.go
@@ -8,5 +8,5 @@
 package libtorch
 
 // #cgo CXXFLAGS: -DGODL_BUILD_CUDA=1 -I/usr/local/cuda/include
-// #cgo LDFLAGS: -Wl,--no-as-needed -ltorch_cuda -lc10_cuda
+// #cgo LDFLAGS: -Wl,--no-as-needed -ltorch_cuda -lc10_cuda -L/usr/local/cuda/lib64 -lcudart
 import "C"